backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
celery[redis]==5.3.6
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
groq==0.9.0
redis==5.0.4
